[CLEANUP] Move the Register to a single class
Move the SmartObject Register to a single class. Also remove the
extension key in the register, so the register of the SmartObjects
is much faster. There are only two places where the extension key
is needed. This places are not in the critical performance path,
so the extension key is calculated on the fly at the situations.

Change-Id: Ia8ce5c2e0b439191d476f1fd44988cd8de1b2517
Reviewed-on: https://review.typo3.org/31410
Reviewed-by: Tim Spiekerkötter
Tested-by: Tim Spiekerkötter